/!\ BLACK to RED Outline.shader!
I found a very efficient way (https://www.youtube.com/watch?v=zpIjme5Ah7Q) to change the outline color of the Alien sprite. Of course, We'll have to use the black color on the animated sprite only for the outline so that we can change it in the game as needed.