Implemented dash talent.
The dash trail is in global coordinates, so it has to be in the main scene,
not in the alien scene. As a result, I had to redistribute responsibilities
between the main scene and the alien a bit: the main scene triggers talents
while the alien scene takes care of lower-level stuff specific to the alien
(e.g. alien animations, acceleration).