Wet and dry footsteps for the alien.
Instead of setting the audio streams directly from the AnimationPlayer's Run
animations, the Run animations now call a method play_step(), which plays a
dry or wet sound as appropriate.

For now we detect collisions between vomit's Poison_Colliders and the
alien's Hit_Collider, because that's easy. But maybe not the best way...