Hack to make footsteps work in Godot editor.
The alien's footstep sounds are animated by Call Method animation tracks.
But those do not work within the Godot editor; they only work when you are
playing the game.

A workaround, so that we can hear footsteps in the editor, is to create a
property which you can animate (with a Property track). Unfortunately, for
that to work, we have to put the "tool" keyword at the top of the alien
script, which is a bit dangerous...