Alien reacts when he is hit (by Binoclard).
If Binoclard hits the alien three times, he turns around and runs away, and
won't start looking for him again for three seconds.

Turned off the Vomit scene's Poison_Collider, which is unnecessary: the
Vomit_Detector tells us when the alien is walking in a puddle, and we don't
want to trigger the Hit animation when he does.